requirements.txt
discord.py
wordcloud
matplotlib